Optimization: use TypeId to index the map
For multithread benchmark this doesn't make much of a difference, however
this seems like a cleaner sollution and also this is slightly more performant
in single thread benchmark.

One downside of using `TypeId::<T>::of` instead of `type_name::<T>` is that
we are now forced to require `T: 'static` (I wished for non-'static TypeId-like
thing for a while, but it doesn't seem possible). While this is a breaking
change, I don't think that restriction will make it any harder to use this
crate.

The only case when this requirement may not be met, is if user has a struct
with a lifetime parameter:
```rust
struct SmtSmtRef<'a> {
    r: &'a SmtSmt,
    _c: Count<Self>, // error
}
```
But this error is easily fixable by just plugging 'static lifetime for all
lifetime parameters:
```rust
struct SmtSmtRef<'a> {
    r: &'a SmtSmt,
    _c: Count<SmtSmtRef<'static>>, // ok
}
```