Optimization: don't use `DashMap::entry` on fast-path
Most of the time an entry for the type is already present (we only need to
create it once). So, using entry api is wasteful as it needs unique access
to the entry, while we don't need it.

This is similar to the optimization of decrementation with the only difference
in that if we can't find an entry with `DashMap::get` we need to fallback to
entry api. This should generally only happen once per type per program run,
so the cost of double lookup negligible.

On my computer this makes multithread benchmark approximatly 3 times faster.