Optimization: use thread local storage
This commit implements the first approach from issue 9.

Every thread now has it's own copy of the hashmap in a thread local storage.
This allows to do most lookups without any need for synchronization (except for
atomic increments that is).

Even though this approach needs more allocations (every `Storage` is wrapped
with `Arc`) and the code is more complicated, it's seems like it's still faster.

On my computer this makes multithread benchmark approximately 2 times faster.